django/requirements.txt
Django
djangorestframework
psycopg2
Pillow
numpy
tensorflow
keras
opencv-python